requirements.txt
flask
sqlalchemy>=1.4
# mysql-connector-python==8.0.33
pymysql
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0

python-dotenv